refactor: move network defaults into the networks service
params/networkdefaults holds the embedded network table and builds the
default network list. It belongs with the service that owns networks,
not in params.

Dissolved into the networks package rather than kept as a subpackage:
BuildDefaultNetworks reads better than defaults.BuildDefaultNetworks,
and the next commit unexports the helpers both halves share.
