+++ pkg/services/networks/default_networks_test.go
@@ -1,4 +1,4 @@
-package networkdefaults
+package networks
 
 import (
 	"strings"
